Refactor user API functions by removing get-users.ts and updating import paths in data-fetching.ts and index.tsx. This change streamlines the codebase and enhances maintainability by consolidating user-related API functions into more specific files.
